// ui/src/lib/types/StructureOutput.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Dimensions } from "./Dimensions";
import type { StructurePanel } from "./StructurePanel";

/**
 * Output shape of a Structure.
 */
export type StructureOutput = "single" | { "paneled": { canvas: Dimensions, panels: Array<StructurePanel>, } };